feat(SPV-1249): move to paymails category
